let's see now

+++ main.py
@@ -163,11 +163,14 @@ def process_link_command(update: Update, context: CallbackContext) -> None:
         upload = anon.upload(file, progressbar=False)
         url = upload.url.geturl()
         update.message.reply_text(f"Here's your download link:\n{url}")
-    except:
+    except Exception as e:
         update.message.reply_text('For some reason i could not download the book you requested.')
+        print({e})
     finally:
-        shutil.rmtree(book.temp_path)
-        shutil.rmtree(pdfpath)
+        if os.path.exists(book.temp_path):
+            shutil.rmtree(book.temp_path)
+        if os.path.exists(pdfpath):
+            shutil.rmtree(pdfpath)
